+++ src/test/converter2/issues/gh2478.ts
@@ -1,0 +1,10 @@
+declare const test: (options?: test.Options) => void;
+
+declare namespace test {
+    interface Options {
+        a: string;
+        b: number;
+    }
+}
+
+export { test };