+++ src/test/issues.c2.test.ts
@@ -1361,4 +1361,15 @@ describe("Issue Tests", () => {
 
         equal(project.children[0].children?.map((c) => c.name), ["Options"]);
     });
+
+    it("Creates a separate namespace for `declare namespace` case with variables #2478", () => {
+        const project = convert();
+
+        equal(project.children?.map((c) => [c.name, c.kind]), [
+            ["test", ReflectionKind.Namespace],
+            ["test", ReflectionKind.Function],
+        ]);
+
+        equal(project.children[0].children?.map((c) => c.name), ["Options"]);
+    });
 });